`showGraph` property in the IState interface defined in App.tsx. It should be of the type`boolean`
when the user clicks "Start Streaming Data". That means you should set the
`showGraph` property of the App's state to `false` in the constructor

To ensure that the graph doesn’t render until a user clicks the ‘Start
Streaming’ button, you should also edit the `renderGraph` method of the
App, adding a condition to render the graph when the application state’s
`showGraph` property is `true`.

modify the `getDataFromServer`method to contact
the server and get data from it continuously instead of just getting data from it
once every time you click the button. Javascript has a way to do things in intervals via the setInterval function.
We can make it continuous with a guard value to stop the interval process
we started.

First, you must enable the `PerspectiveViewerElement` to behave like an
HTMLElement. To do this, you can extend the `HTMLElement` class from the
`PerspectiveViewerElement` interface

modify the `componentDidMount` method.Since you changed the `PerspectiveViewerElement` to extend the
`HTMLElement` earlier, you can simplify the `const elem` definition by
assigning it directly to the result of `document.getElementsByTagName`.

Finally, you need to add more attributes to the element. You’ll need to add
the following attributes: `view`, `column-pivots`, `row-pivots`, `columns`
and `aggregates`.